move fixtures to /test/conftest.py
