+++ api/conftest.py
@@ -8,16 +8,10 @@ from django.core.cache import cache
 from flag_engine.segments.constants import EQUAL
 from moto import mock_dynamodb
 from mypy_boto3_dynamodb.service_resource import DynamoDBServiceResource, Table
-from pytest_django.fixtures import SettingsWrapper
 from rest_framework.authtoken.models import Token
 from rest_framework.test import APIClient
 
 from api_keys.models import MasterAPIKey
-from environments.dynamodb.dynamodb_wrapper import (
-    DynamoEnvironmentV2Wrapper,
-    DynamoEnvironmentWrapper,
-    DynamoIdentityWrapper,
-)
 from environments.identities.models import Identity
 from environments.identities.traits.models import Trait
 from environments.models import Environment, EnvironmentAPIKey
@@ -60,7 +54,6 @@ from projects.tags.models import Tag
 from segments.models import Condition, Segment, SegmentRule
 from task_processor.task_run_method import TaskRunMethod
 from users.models import FFAdminUser, UserPermissionGroup
-from util.mappers import map_environment_to_environment_document
 
 trait_key = "key1"
 trait_value = "value1"
@@ -619,83 +612,3 @@ def flagsmith_environments_v2_table(dynamodb: DynamoDBServiceResource) -> Table:
         ],
         BillingMode="PAY_PER_REQUEST",
     )
-
-
-@pytest.fixture()
-def flagsmith_environment_api_key_table(dynamodb: "DynamoDBServiceResource") -> "Table":
-    return dynamodb.create_table(
-        TableName="flagsmith_environment_api_key",
-        KeySchema=[{"AttributeName": "key", "KeyType": "HASH"}],
-        AttributeDefinitions=[
-            {"AttributeName": "key", "AttributeType": "S"},
-        ],
-        BillingMode="PAY_PER_REQUEST",
-    )
-
-
-@pytest.fixture()
-def flagsmith_environment_table(dynamodb: "DynamoDBServiceResource") -> "Table":
-    return dynamodb.create_table(
-        TableName="flagsmith_environments",
-        KeySchema=[{"AttributeName": "api_key", "KeyType": "HASH"}],
-        AttributeDefinitions=[
-            {"AttributeName": "api_key", "AttributeType": "S"},
-        ],
-        BillingMode="PAY_PER_REQUEST",
-    )
-
-
-@pytest.fixture()
-def flagsmith_project_metadata_table(dynamodb: "DynamoDBServiceResource") -> "Table":
-    return dynamodb.create_table(
-        TableName="flagsmith_project_metadata",
-        KeySchema=[{"AttributeName": "id", "KeyType": "HASH"}],
-        AttributeDefinitions=[
-            {"AttributeName": "id", "AttributeType": "N"},
-        ],
-        BillingMode="PAY_PER_REQUEST",
-    )
-
-
-@pytest.fixture
-def dynamodb_identity_wrapper(
-    settings: SettingsWrapper,
-    flagsmith_identities_table: Table,
-) -> DynamoIdentityWrapper:
-    settings.IDENTITIES_TABLE_NAME_DYNAMO = flagsmith_identities_table.name
-    return DynamoIdentityWrapper()
-
-
-@pytest.fixture
-def dynamodb_wrapper_v2(
-    settings: SettingsWrapper,
-    flagsmith_environments_v2_table: Table,
-) -> DynamoEnvironmentV2Wrapper:
-    settings.ENVIRONMENTS_V2_TABLE_NAME_DYNAMO = flagsmith_environments_v2_table.name
-    return DynamoEnvironmentV2Wrapper()
-
-
-@pytest.fixture()
-def dynamo_enabled_project_environment_one_document(
-    flagsmith_environment_table: Table,
-    dynamo_enabled_project_environment_one: Environment,
-) -> dict:
-    environment_dict = map_environment_to_environment_document(
-        dynamo_enabled_project_environment_one
-    )
-
-    flagsmith_environment_table.put_item(
-        Item=environment_dict,
-    )
-    return environment_dict
-
-
-@pytest.fixture()
-def dynamo_environment_wrapper(
-    flagsmith_environment_table: Table,
-    settings: SettingsWrapper,
-) -> DynamoEnvironmentWrapper:
-    wrapper = DynamoEnvironmentWrapper()
-
-    wrapper.table_name = flagsmith_environment_table.name
-    return wrapper
